:art: refactor: 코드 분리 및 리팩토링
