+++ src/main/java/com/educator/eduo/security/TokenProvider.java
@@ -1,5 +1,6 @@
 package com.educator.eduo.security;
 
+import com.educator.eduo.auth.model.dto.JwtResponse;
 import com.educator.eduo.auth.model.entity.Token;
 import com.educator.eduo.auth.model.entity.User;
 import io.jsonwebtoken.*;
@@ -85,6 +86,18 @@ public class TokenProvider implements InitializingBean {
         }
 
         return false;
+    }
+
+    public JwtResponse createJwtResponse(Authentication authentication, Token token) {
+        User user = (User) authentication.getPrincipal();
+
+        return JwtResponse.builder()
+                .accessToken(token.getAccessToken())
+                .refreshToken(token.getRefreshToken())
+                .userId(user.getUserId())
+                .name(user.getName())
+                .role(user.getRole())
+                .build();
     }
 
     public String resolveToken(HttpServletRequest request) {
